Move ChildSize hash to Spectral
